+++ mmd_uuunyaa_tools/thumbnails/PHYSICS_CLOTH_NOTHING.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:22d84b3602c00a5e27fbf9e7f0f9a4fbd1f8f3dc52f5844e1c4f2ac72061bf2c
+size 12127
