+++ mmd_uuunyaa_tools/thumbnails/PHYSICS_COLLISION_NOTHING.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:06ab142c3bad172a564f50931dd667c07af918ed7298107b0c41f79cd761cc8f
+size 143
