+++ mmd_uuunyaa_tools/thumbnails/PHYSICS_COLLISION_THIN.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4a9efb891302094ce8e7319d4b17a2bbfa8321c5cb431a57a1221ea4993162ae
+size 200
